envoy: move all methods from xDSServer into xds_server.go
Currently, `pkg/envoy/ciliumenvoyconfig.go` contains parts of the
`xdsServer` (implementing functions related to `resources`) and the
"parse logic" that is used when parsing the `CiliumEnvoyConfig`.

In preparation to move the parse logic into package `pkg/ciliumenvoyconfig`,
this commit moves the `resources` and the following implementing methods of
`xdsServer` from `pkg/envoy/ciliumenvoyconfig.go` into `pkg/envoy/xds_server.go`.

* `UpsertEnvoyResources`
* `UpdateEnvoyResources`
* `DeleteEnvoyResources`

Note: No changes were made at the actual implementations.
